blocking/requirements.txt
faust-streaming
aiokafka==0.12.0